refactor: ProjectGrid 컴포넌트를 새로운 위치로 이동 및 관련 경로 수정
